add update and delete

+++ controllers/contactController.js
@@ -6,7 +6,7 @@ const Contact = require("../models/contactModel");
 //@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find({user_id: req.user.id});
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -25,7 +25,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
-    user_id: req.user.id
+    user_id: req.user.id,
   });
   res.status(200).json(contact);
 });
@@ -53,6 +53,12 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contact");
+  }
+
   const updateContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -71,6 +77,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contact");
+  }
+
   // await Contact.remove()
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
